requirements.txt
torch
torchvision
efficientnet_pytorch
numpy
pandas
scipy
scikit-learn
matplotlib
tqdm
pillow